requirements.txt
Cython==0.29.24
PyYAML==5.4.1